Fixed config for sudo to unprivileged users
http://docs.ansible.com/ansible/latest/become.html#becoming-
an-unprivileged-user

Newer versions of Ansible changed how commands are run.  In
modern Ansible, if you want to become an unprivileged user,
that user will not have permission to write files to /tmp.
You must change the ssh mode to "pipelining" the commands.
The default is to copy the commands as a text file, then
execute the file.  This results in secret stuff sitting on
the hard drive for all to see.